Can walk up and down the directory tree.
At login, the path logged in to is saved in an object along with the
connection session. When selected a new directory, that directory is
added to the original connection directory and the contents of the
rsulting directory are listed. If the .. directory is selected, one node
is removed from the saved directory and the contents of the resulting
directory are listed.