Loading Connection content is done in the ConnectionActivity.
The MainActivity only initiates Connections. When a Connection is
initiated, the Connection Activity is started with reference to the
Connection that was just initiated. The Connection Activity is then in
charge of displaying the content of the Connection of interest (and in
the future, providing the option of switching to other active
connections.