Files saved when backed out of.
Commands were also refactored. They are now Integer constants within the
Connection Service.

Also refactored command thread to take command,
parameter and content instead of "path" and whatever else it was
exploding to be. The parameter is any parameter the JSCH SFTPChannel
requires for the specific command, which will often be a path.

The CD command has been added and is called whenever a directory list
item is clicked. The connection session path is now updated only when
the CD command returns. This prevents adjusting the path when a file is
opened, therefore allowing the file to be reopened after it is backed
out of.